Fixed code smells

+++ src/main/java/nl/fontys/s3/rentride_be/business/useCases/payment/CreatePaymentSessionUseCase.java
@@ -3,5 +3,5 @@ package nl.fontys.s3.rentride_be.business.useCases.payment;
 import com.stripe.exception.StripeException;
 
 public interface CreatePaymentSessionUseCase {
-    String createPaymentSession(String description, Long price) throws StripeException;
+    String createPaymentSession(String description, double price) throws StripeException;
 }
